+++ freediscovery/categorization.py
@@ -105,7 +105,7 @@ class NearestNeighborRanker(BaseEstimator, RankerMixin):
 
     """
 
-    def __init__(self, n_neighbors=1, radius=1.0,
+    def __init__(self, radius=1.0,
                  algorithm='ball_tree', leaf_size=30, n_jobs=1, **kwargs):
 
         # define nearest neighbors search objects for positive and negative samples