+++ freediscovery/text.py
@@ -13,6 +13,7 @@ from sklearn.externals import joblib
 from sklearn.externals.joblib import Parallel, delayed
 from sklearn.feature_extraction.text import TfidfTransformer, TfidfVectorizer
 from sklearn.preprocessing import normalize
+from sklearn.pipeline import make_pipeline
 
 from .pipeline import PipelineFinder
 from .utils import generate_uuid, _rename_main_thread
@@ -43,7 +44,7 @@ def _vectorize_chunk(dsid_dir, k, pars, pretend=False):
     fe = HashingVectorizer(input='filename', norm=None, decode_error='ignore',
            non_negative=True, **hash_opts) 
     if pretend:
-        return
+        return fe
     fset_new = fe.transform(filenames[mslice])
 
     fset_new.eliminate_zeros()
@@ -176,6 +177,7 @@ class _BaseTextTransformer(object):
         """ Number of documents in the dataset """
         return len(self._pars['filenames'])
 
+
     def list_datasets(self):
         """ List all datasets in the working directory """
         import traceback
@@ -227,12 +229,59 @@ class FeatureVectorizer(_BaseTextTransformer):
 
     _wrapper_type = "vectorizer"
 
-    def preprocess(self, data_dir, file_pattern='.*', dir_pattern='.*',  n_features=11000000,
+    def preprocess(self, data_dir, file_pattern='.*', dir_pattern='.*',  n_features=None,
             chunk_size=5000, analyzer='word', ngram_range=(1, 1), stop_words='None',
-            n_jobs=1, use_idf=False, sublinear_tf=True, binary=False, use_hashing=True,
+            n_jobs=1, use_idf=False, sublinear_tf=True, binary=False, use_hashing=False,
             norm='l2', min_df=0.0, max_df=1.0):
-        """Initalize the features extraction. See sklearn.feature_extraction.text for a
-        detailed description of the input parameters """
+        """Initalize the features extraction.
+
+        See sklearn.feature_extraction.text for a detailed description of the input parameters
+
+        Parameters
+        ----------
+
+        analyzer : string, {'word', 'char'} or callable
+            Whether the feature should be made of word or character n-grams.
+            If a callable is passed it is used to extract the sequence of features
+            out of the raw, unprocessed input.
+        ngram_range : tuple (min_n, max_n)
+            The lower and upper boundary of the range of n-values for different
+            n-grams to be extracted. All values of n such that min_n <= n <= max_n
+            will be used.
+        stop_words : string {'english'}, list, or None (default)
+            If a string, it is passed to _check_stop_list and the appropriate stop
+            list is returned. 'english' is currently the only supported string
+            value.
+            If None, no stop words will be used. max_df can be set to a value
+            in the range [0.7, 1.0) to automatically detect and filter stop
+            words based on intra corpus document frequency of terms.
+        max_df : float in range [0.0, 1.0] or int, default=1.0
+            When building the vocabulary ignore terms that have a document
+            frequency strictly higher than the given threshold (corpus-specific
+            stop words).
+            If float, the parameter represents a proportion of documents, integer
+            absolute counts.
+            This parameter is ignored if vocabulary is not None.
+        min_df : float in range [0.0, 1.0] or int, default=1
+            When building the vocabulary ignore terms that have a document
+            frequency strictly lower than the given threshold. This value is also
+            called cut-off in the literature.
+            If float, the parameter represents a proportion of documents, integer
+            absolute counts.
+            This parameter is ignored if vocabulary is not None.
+        max_features : int or None, default=None or 100001
+            If not None, build a vocabulary that only consider the top
+            max_features ordered by term frequency across the corpus.
+        binary : boolean, default=False
+            If True, all non-zero term counts are set to 1. This does not mean
+            outputs will have only 0/1 values, only that the tf term in tf-idf
+            is binary. (Set idf and normalization to False to get 0/1 outputs.)
+        norm : 'l1', 'l2' or None, optional
+            Norm used to normalize term vectors. None for no normalization.
+        use_idf : boolean, default=True
+            Enable inverse-document-frequency reweighting.
+
+        """
         data_dir = os.path.normpath(data_dir)
 
         if not os.path.exists(data_dir):
@@ -251,6 +300,9 @@ class FeatureVectorizer(_BaseTextTransformer):
             raise WrongParameter('len(gram_range=={}!=2'.format(len(ngram_range)))
         if stop_words not in ['None', 'english', 'english_alphanumeric']:
             raise WrongParameter('stop_words')
+
+        if n_features is None and use_hashing:
+            n_features = 100001 # default size of the hashing table
 
         filenames_rel = [os.path.relpath(el, data_dir) for el in filenames]
         self.dsid = dsid = generate_uuid()
@@ -314,7 +366,7 @@ class FeatureVectorizer(_BaseTextTransformer):
         if use_hashing:
             # just make sure that we can initialize the vectorizer
             # (easier outside of the paralel loop
-            _vectorize_chunk(dsid_dir, 0, pars, pretend=True)
+            vect = _vectorize_chunk(dsid_dir, 0, pars, pretend=True)
 
         processing_lock =  os.path.join(dsid_dir, 'processing')
         _touch(processing_lock)
@@ -332,7 +384,8 @@ class FeatureVectorizer(_BaseTextTransformer):
                     tfidf = TfidfTransformer(norm=pars['norm'], use_idf=True,
                                               sublinear_tf=pars['sublinear_tf'])
                     res = tfidf.fit_transform(res)
-                self.vect = None
+                    vect = make_pipeline(vect, tfidf)
+                self.vect = vect
             else:
                 opts_tfidf = {key: val for key, val in pars.items() \
                         if key in ['stop_words', 'use_idf', 'ngram_range', 'analyzer',
@@ -343,8 +396,8 @@ class FeatureVectorizer(_BaseTextTransformer):
                             norm=pars['norm'],
                             decode_error='ignore', **opts_tfidf)
                 res = tfidf.fit_transform(pars['filenames_abs'])
-                joblib.dump(tfidf, os.path.join(dsid_dir, 'vectorizer'))
                 self.vect = tfidf
+            joblib.dump(self.vect, os.path.join(dsid_dir, 'vectorizer'))
 
             if pars['norm'] is not None:
                 res = normalize(res, norm=pars['norm'], copy=False)
@@ -387,6 +440,20 @@ class FeatureVectorizer(_BaseTextTransformer):
             out.append(terms[idx])
         return out
 
+    @property
+    def n_features_(self):
+        """ Number of features of the vecotorizer"""
+        from sklearn.feature_extraction.text import HashingVectorizer
+        from sklearn.pipeline import Pipeline
+        vect = self._load_model()
+        if hasattr(vect, 'vocabulary_'):
+            return len(vect.vocabulary_)
+        elif isinstance(vect, HashingVectorizer):
+            return vect.get_params()['n_features']
+        elif isinstance(vect, Pipeline):
+            return vect.named_steps['hashingvectorizer'].get_params()['n_features']
+        else:
+            raise ValueError
 
 
     def _aggregate_features(self):